requirements.txt
pandas==2.2.3
sentence-transformers==4.0.2
faiss-cpu==1.10.0
llama-cpp-python==0.3.8
flask==3.1.0